Move mock backup and restore collections to data
