+++ src/internal/m365/mock/connector.go
@@ -3,9 +3,13 @@ package mock
 import (
 	"context"
 
+	"github.com/alcionai/clues"
+
 	"github.com/alcionai/corso/src/internal/common/idname"
 	"github.com/alcionai/corso/src/internal/common/prefixmatcher"
 	"github.com/alcionai/corso/src/internal/data"
+	"github.com/alcionai/corso/src/internal/kopia"
+	kinject "github.com/alcionai/corso/src/internal/kopia/inject"
 	"github.com/alcionai/corso/src/internal/operations/inject"
 	"github.com/alcionai/corso/src/pkg/backup/details"
 	"github.com/alcionai/corso/src/pkg/control"
@@ -44,6 +48,16 @@ func (ctrl Controller) ProduceBackupCollections(
 	error,
 ) {
 	return ctrl.Collections, ctrl.Exclude, ctrl.Err == nil, ctrl.Err
+}
+
+func (ctrl *Controller) CollectMetadata(
+	ctx context.Context,
+	r kinject.RestoreProducer,
+	man kopia.ManifestEntry,
+	tenantID string,
+	errs *fault.Bus,
+) ([]data.RestoreCollection, error) {
+	return nil, clues.New("not implemented")
 }
 
 func (ctrl Controller) IsServiceEnabled(
